Implement Validator Error and Exception Handling

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,24 +1,35 @@
-﻿using MediatR;
+﻿using Application.Core;
+using MediatR;
 using Persistence;
 
 namespace Application.Activities.Commands
 {
     public class DeleteActivity
     {
-        public class Command: IRequest
+        public class Command: IRequest<Result<Unit>>
         {
             public required string Id { get; set; }
         }
 
-        public class Handler(AppDbContext dbContext) : IRequestHandler<Command>
+        public class Handler(AppDbContext dbContext) : IRequestHandler<Command, Result<Unit>>
         {
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
-                var activity = await dbContext.Activities.FindAsync([request.Id], cancellationToken)
-                    ?? throw new Exception("Not found activity");
+                var activity = await dbContext.Activities.FindAsync([request.Id], cancellationToken);
+                if (activity == null)
+                {
+                    return Result<Unit>.Failure("Activity not found", 404);
+                }
+
 
                 dbContext.Activities.Remove(activity);
-                await dbContext.SaveChangesAsync(cancellationToken);
+                var result = await dbContext.SaveChangesAsync(cancellationToken) > 0;
+                if (!result)
+                {
+                    return Result<Unit>.Failure("Failed to delete activity", 400);
+                }
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
